Phase 3.3: Split SyncService into detector + downloader + uploader
Extract from sync_service.dart (1,009 → ~160 LOC):
- sync_models.dart (~103 LOC): FileChangeRecord, SyncCheckResponse
- sync_change_detector.dart (~234 LOC): remote/local change queries
- sync_downloader.dart (~374 LOC): server→local entity sync + asset download
- sync_uploader.dart (~217 LOC): local→server entity upload

SyncService is now a thin orchestrator that delegates to 3 services.
Re-exports sync_models.dart so existing consumers need no import changes.
All Flutter tests pass. Zero analyze errors.
